pyproject.toml
[project]
name = "fintech_ux_research"
version = "0.1.0"
description = "fintech_ux_research using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[anthropic,google-genai,groq,litellm,tools]==1.14.2",
    "markdown>=3.10.2",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "streamlit>=1.56.0",
]

[project.scripts]
fintech_ux_research = "fintech_ux_research.main:run"
run_crew = "fintech_ux_research.main:run"
train = "fintech_ux_research.main:train"
replay = "fintech_ux_research.main:replay"
test = "fintech_ux_research.main:test"
run_with_trigger = "fintech_ux_research.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
